refactor(form): apply changes proposed in CR
