+++ libs/shared/src/components/form-field/index.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import { Edit } from "@mui/icons-material";
-import { ChangeEvent, ChangeEventHandler } from "react";
+import { ChangeEvent } from "react";
 import { FieldValues, useController, UseControllerProps, UseFormStateReturn } from "react-hook-form";
 
 import { FormFieldType } from "../../types";
@@ -9,17 +9,17 @@ import { renderHelperText } from "./render-helper-text";
 import * as Styled from "./styled";
 
 export interface FormFieldProps extends Pick<UseControllerProps, "name" | "defaultValue" | "control"> {
-  fieldType: FormFieldType;
+  type: FormFieldType;
   variant?: FormTextFieldVariant;
   rows?: number;
   label?: string;
   helperText?: string;
   hideEditIcon?: boolean;
-  onChange?: (event: ChangeEvent) => void;
+  onChange?: () => void;
 }
 
 export const FormField: React.FC<FormFieldProps> = ({
-  fieldType,
+  type,
   name,
   defaultValue,
   control,
@@ -31,22 +31,23 @@ export const FormField: React.FC<FormFieldProps> = ({
   onChange
 }: FormFieldProps) => {
   const {
-    field: { name: fieldName, value, onChange: controllerOnChange },
+    field: { name: fieldName, value, onChange: OnFieldChange },
     formState: { errors }
   } = useController({ name, defaultValue, control });
 
   const fieldErrors: UseFormStateReturn<FieldValues>["errors"] = errors[name];
 
-  const handleChange: ChangeEventHandler = event => {
-    controllerOnChange(event);
+  const handleChange = (event: ChangeEvent) => {
+    OnFieldChange(event);
+
     if (onChange) {
-      onChange(event);
+      onChange();
     }
   };
 
   return (
     <Styled.Field>
-      {renderField({ type: fieldType, name: fieldName, value, handleChange, variant, errors: fieldErrors, label, rows })}
+      {renderField({ type, name: fieldName, value, handleChange, variant, errors: fieldErrors, label, rows })}
       {!hideEditIcon && <Edit />}
       {renderHelperText(fieldErrors, helperText)}
     </Styled.Field>
